Adding a name function to the program
In this branch, I began working on the name function of this program.  This section of the program will ask the user to provide their name, but will have various parameters that it will expect the user to honor.  For example: the program will reject answers that have symbols in them, such as @, $, etc.  The program will also reject answers that have lengths that are longer than 100 characters.

These are initial parameters that I want to include in the program because they are reasonable restraints.  Without them, users would be able to enter in random sequences of symbols without a length limit.  Having such parameters is necessary to have reasonable limits on what users can enter for their answers.