"Are you sure?" example
This branch will add an "are you sure? function to the program, regarding the age variable.  After the user answers the Interrogator's initial round of questions, Interrogator stores those answers in a list variable called "main_memory".  Interrogator then reviews the answers contained within said list variable and asks the user whether all of the information is correct.

This example shows how the user will be able to answer "age" to Interrogator's 2nd round question.  When the user answers "age" to the Interrogator's question, Interrogator then prompts the user with several follow-up questions in order to make sure that the user's answer for the age variable is accurate.